ai-korean-be/asr_service/requirements.txt
# === Core Numeric / Audio ===
numpy>=1.24
scipy>=1.11
torch>=2.2          # dùng bản CUDA tương ứng môi trường (pip/whl riêng)
torchaudio>=2.2     # khớp phiên bản torch
soundfile>=0.12     # cần libsndfile của hệ thống
ffmpeg-python>=0.2  # cần ffmpeg hệ thống

# === Transformers / HF ===
transformers>=4.41
huggingface-hub>=0.23
accelerate>=0.33    # khuyến nghị để device map/loading êm

# === (Tuỳ chọn) Whisper backend ===
openai-whisper>=20231117
# hoặc: faster-whisper>=1.0.0  # nếu muốn RT yếu tài nguyên

# === FastAPI server ===
fastapi>=0.112
uvicorn[standard]>=0.30
python-multipart>=0.0.9
pydantic>=2.8
anyio>=4.4

# === Audio convenience ===
pydub>=0.25

# === TextGrid (tuỳ chọn cho priors) ===
praat-textgrids>=1.4.0      # 'textgrid' package
tgt>=1.5.0                  # lib thay thế đọc TextGrid

# === Utils ===
requests>=2.32
phonemizer>=3.3    # (tuỳ chọn) nếu sau này cần phoneme conversion
